+++ TelegramBotAPI-extensions-api/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/extensions/api/stickers/CreateNewStaticStickerSet.kt
@@ -3,12 +3,12 @@ package com.github.insanusmokrassar.TelegramBotAPI.extensions.api.stickers
 import com.github.insanusmokrassar.TelegramBotAPI.bot.RequestsExecutor
 import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.FileId
 import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.MultipartFile
-import com.github.insanusmokrassar.TelegramBotAPI.requests.stickers.CreateNewStickerSet
+import com.github.insanusmokrassar.TelegramBotAPI.requests.stickers.CreateNewStaticStickerSet
 import com.github.insanusmokrassar.TelegramBotAPI.types.CommonUser
 import com.github.insanusmokrassar.TelegramBotAPI.types.UserId
 import com.github.insanusmokrassar.TelegramBotAPI.types.stickers.MaskPosition
 
-suspend fun RequestsExecutor.createNewStickerSet(
+suspend fun RequestsExecutor.createNewStaticStickerSet(
     userId: UserId,
     name: String,
     sticker: FileId,
@@ -16,10 +16,10 @@ suspend fun RequestsExecutor.createNewStickerSet(
     containsMasks: Boolean? = null,
     maskPosition: MaskPosition? = null
 ) = execute(
-    CreateNewStickerSet(userId, name, sticker, emojis, containsMasks, maskPosition)
+    CreateNewStaticStickerSet(userId, name, sticker, emojis, containsMasks, maskPosition)
 )
 
-suspend fun RequestsExecutor.createNewStickerSet(
+suspend fun RequestsExecutor.createNewStaticStickerSet(
     userId: UserId,
     name: String,
     sticker: MultipartFile,
@@ -27,28 +27,28 @@ suspend fun RequestsExecutor.createNewStickerSet(
     containsMasks: Boolean? = null,
     maskPosition: MaskPosition? = null
 ) = execute(
-    CreateNewStickerSet(userId, name, sticker, emojis, containsMasks, maskPosition)
+    CreateNewStaticStickerSet(userId, name, sticker, emojis, containsMasks, maskPosition)
 )
 
 
-suspend fun RequestsExecutor.createNewStickerSet(
+suspend fun RequestsExecutor.createNewStaticStickerSet(
     user: CommonUser,
     name: String,
     sticker: FileId,
     emojis: String,
     containsMasks: Boolean? = null,
     maskPosition: MaskPosition? = null
-) = createNewStickerSet(
+) = createNewStaticStickerSet(
     user.id, name, sticker, emojis, containsMasks, maskPosition
 )
 
-suspend fun RequestsExecutor.createNewStickerSet(
+suspend fun RequestsExecutor.createNewStaticStickerSet(
     user: CommonUser,
     name: String,
     sticker: MultipartFile,
     emojis: String,
     containsMasks: Boolean? = null,
     maskPosition: MaskPosition? = null
-) = createNewStickerSet(
+) = createNewStaticStickerSet(
     user.id, name, sticker, emojis, containsMasks, maskPosition
 )